Updated oversight in Battleship.cpp
Fixed an instance in the code where if the user had inputted coordinates that was a previous hit spot by the computer, it would not run anything in the terminal. I simply added an or condition to the Else If statement including that if the computer board had '~' in the place of where the user hits that it would still run that the user has not hit any ships.